requirements.txt
numpy==1.26.4
scipy==1.11.4
matplotlib
spatialmath-python
roboticstoolbox-python